Fixed overzealous recursion prevention. Closes #78.
In order to track which page parts are currently being rendered, we
need a data structure that's shared across the tag rendering
stack. But that means we can't rely on tag.locals to pop each page
part off the stack when we finish rendering it. So instead we use
tag.globals, and manually pop the rendering stack.